Added event handler check for custom errors
If the underlying code (that's implementing OAuth2orize) throws an error, OAuth2orize will intercept this error and relay it back to the client. This is good for OAuth errors, but if the error is a bug in the underlying code, all means of debugging/logging is lost. For this reason I've added a check to see if the error is an OAuth error. If it's not, it'll be passed on to the next middleware.