Add handleAllErrors flag to errorHandler's options
The errorHandler middleware "swallows" unknown errors in direct mode by
setting the status code to 500 and sending the response. This makes it
implossible for a user to implement his own global error handler, for
example to print the stack traces of all uncaught errors (think about
errors with the database).

This change adds `handleAllErrors` options flag with the default value `true`:

- When set to `true`, errorHandler will behave as before
- When set to `false, errorHandler will call `next(err)` for unknown
  errors and thus propagate the error to the global error handler